Subdirectory for library modules
