+++ __init__.py
@@ -1,3 +1,5 @@
-from .path_checker import PathChecker,\
-	make_default_file_name, make_default_path
-from .arg_path_checker import ArgPathChecker
+from .path_checker import\
+	ArgPathChecker,\
+	make_default_file_name,\
+	make_default_path,\
+	PathChecker
